+++ src/DefaultBuilder/test/Microsoft.AspNetCore.NativeAotTests/Microsoft.AspNetCore.NativeAotTests.proj
@@ -1,9 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <ItemGroup>
-    <TestConsoleAppSourceFiles Include="SlimBuilderDoesNotDependOnX509Test.cs">
-      <DisabledFeatureSwitches>System.Text.Json.JsonSerializer.IsReflectionEnabledByDefault</DisabledFeatureSwitches>
-    </TestConsoleAppSourceFiles>
     <TestConsoleAppSourceFiles Include="UseStartupThrowsForStructContainersTest.cs" />
   </ItemGroup>
